Update README-DE.md changed Links and more
I've updatet the URLs where possible to the German part of the side. Ive changed some minor grammatical or Understanding Questions. 
I've inclined the Link to a Spanish side. I did not delete it, because it's a neat self-explaining slideshow.
Changed the Link to Gitkraken to GithubDesktop, a not bloated client, without premium functionalities and proper documentation, like https://github.com/desktop/desktop/blob/development/docs/process/what-is-desktop.md